Introduce DefaultHtmlBuilder
to have a central and easy to use entry point for configuring the HTML
rendering, and to hide the details of FlatHtml and IndentedHtml.

Adds static methods in Indenter for easy changing the indent depth.

This commits also adds @Deprecated annotations to some methods and
fields that should be removed in a future version of the library, namely
 - some static fields of Config
 - the public visibility of FlatHtml and IndentedHtml